+++ R/f_analysis_multiarm_means.R
@@ -13,9 +13,9 @@
 ## |
 ## |  Contact us for information about our services: info@rpact.com
 ## |
-## |  File version: $Revision: 8276 $
-## |  Last changed: $Date: 2024-09-26 13:37:54 +0200 (Do, 26 Sep 2024) $
-## |  Last changed by: $Author: pahlke $
+## |  File version: $Revision: 8545 $
+## |  Last changed: $Date: 2025-02-10 12:37:59 +0100 (Mo, 10 Feb 2025) $
+## |  Last changed by: $Author: wassmer $
 ## |
 
 #' @include f_logger.R
@@ -514,8 +514,8 @@ NULL
             if (isFALSE(directionUpper)) {
                 separatePValues[treatmentArm, k] <- 1 - separatePValues[treatmentArm, k]
                 overallPValues[treatmentArm, k] <- 1 - overallPValues[treatmentArm, k]
-                # testStatistics[treatmentArm, k] <- -testStatistics[treatmentArm, k]
-                # overallTestStatistics[treatmentArm, k] <- -overallTestStatistics[treatmentArm, k]
+#                testStatistics[treatmentArm, k] <- -testStatistics[treatmentArm, k]
+#                overallTestStatistics[treatmentArm, k] <- -overallTestStatistics[treatmentArm, k]
             }
         }
     }
@@ -523,7 +523,7 @@ NULL
     .setWeightsToStageResults(design, stageResults)
 
     # Calculation of single stage adjusted p-Values and overall test statistics
-    # for determination of RCIs
+    # for determination of RCIs for combination tests
     if (calculateSingleStepAdjusted) {
         singleStepAdjustedPValues <- matrix(NA_real_, nrow = gMax, ncol = kMax)
         combInverseNormal <- matrix(NA_real_, nrow = gMax, ncol = kMax)
@@ -811,6 +811,7 @@ NULL
                 thetaUp <- thetaUpLimit
                 iteration <- 30
                 prec <- 1
+                
                 while (prec > tolerance) {
                     theta <- (thetaLow + thetaUp) / 2
                     stageResults <- .getStageResultsMeansMultiArm(
@@ -819,6 +820,7 @@ NULL
                         intersectionTest = intersectionTest, normalApproximation = normalApproximation,
                         varianceOption = varianceOption, calculateSingleStepAdjusted = FALSE
                     )
+
                     conditionalDunnettSingleStepRejected <- .getConditionalDunnettTestForCI(
                         design = design, stageResults = stageResults, treatmentArm = treatmentArm
                     )